+++ tests/threadHDF5.cpp
@@ -98,12 +98,12 @@ BOOST_AUTO_TEST_CASE(test_parallel_acess_of_morphology)
     boost::filesystem::path path(BBP_TESTDATA);
     path /= "circuitBuilding_1000neurons/morphologies/h5/C040426.h5";
 
-    const brion::Morphology morphology(path.string());
+    brion::Morphology morphology(path.string());
 #pragma omp parallel
     {
-        const brion::Vector4fsPtr points = morphology.readPoints();
-        const brion::Vector2isPtr sections = morphology.readSections();
-        const brion::SectionTypesPtr types = morphology.readSectionTypes();
+        const auto points = morphology.readPoints();
+        const auto sections = morphology.readSections();
+        const auto types = morphology.readSectionTypes();
         BOOST_CHECK(!points->empty());
         BOOST_CHECK(!sections->empty());
         BOOST_CHECK(!types->empty());